refactor [#84] Custom useQuery 이름에 `~Query` postfix 추가
