+++ src/features/user/query/use-user-query.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { getUser } from "../apis";
 
-export function useUser() {
+export function useUserQuery() {
   const { data } = useQuery({
     queryKey: ["user"],
     queryFn: () => getUser(),
